requirements.txt
streamlit==1.52.2
tensorflow==2.20.0
numpy>=1.26.0,<2.0
pillow==10.4.0
h5py>=3.10.0
